+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This is a changelog for **ZachranObed** application.
 - **ZOB-113** Lock Android for portrait screen orientation in Manifest.
 
 ### Changed
+- **ZOB-296** Update allergens UI in food offer form.
 
 ### Removed
 - **ZOB-295** Remove updating delivery state to NOT_USED from application.